requirements.txt
# APS Alt Text Pipeline — Python Dependencies
# Install with: pip install -r requirements.txt

# PDF parsing
PyMuPDF>=1.24.0

# Image processing
Pillow>=10.0.0

# Excel output
openpyxl>=3.1.0

# Google Gemini Vision LLM
google-generativeai>=0.8.0

# Web server
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9

# Environment variables
python-dotenv>=1.0.0
